Fixed a bug with alignment times
Replaceecodes.m looks for a pair of triggers that are x seconds apart,
where x is the length of the first good trial, and aligns spike2 trigger
times and rex times to that. However, if this fails, it now defaults to
aligning the first trigger with the first rex trial start time. This
might not always yield correct alignments.