fix for no instances of a cluster in a trial
rex_trial_raw used to only create an entry in the cell array spk if
there was at least one spike from that cluster in the trial. This
created issues downstream when filling out the allspkclus cell array.
Changed it so that trials without a particular cluster are filled out as
one nan.